Removed the last version and divided material generator
